% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_estimate_Imm_selected}
\alias{exp_estimate_Imm_selected}
\title{Estimated Causal Effects for Selected Study Sites (data from Naumann et al. (2018))}
\format{
data
}
\description{
Estimated Causal Effects for Selected Study Sites (data from Naumann et al. (2018))
}
\keyword{datasets}
