% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sps_estimator.R
\name{plot.sps_estimator}
\alias{plot.sps_estimator}
\title{Plot function for \code{sps_estimator}}
\usage{
\method{plot}{sps_estimator}(x, ...)
}
\arguments{
\item{x}{Output from function \code{sps_estimator()}}

\item{...}{Other arguments.}
}
\value{
\code{sps_plot_est} plots estimates and 95% confidence intervals for the site-specific ATEs and average-site ATE.
}
\description{
Plot function for \code{sps_estimator}
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
