% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps_cv_new.R
\name{sps_cv_new}
\alias{sps_cv_new}
\title{Site-level Cross-Validation for Synthetic Purposive Sampling}
\usage{
sps_cv_new(
  out = NULL,
  estimates_selected = NULL,
  K = 2,
  max_iter = 100,
  seed = 1234
)
}
\arguments{
\item{out}{Output from function \code{sps()}}

\item{estimates_selected}{data.frame with two columns: the first column represents estimates of the site-specific ATEs for the selected sites and the second column represents its corresponding standard error. The number of rows is equal to the number of the selected sites and \code{rownames(estimates_selected)} should be the names of the selected sites.}

\item{K}{(Default = \code{2}) Fold of the cross-validation.}

\item{max_iter}{(Default = \code{100}) How many times the function repeats K-fold cross-validation.}

\item{seed}{Numeric. \code{seed} used internally. Default = \code{1234}.}
}
\value{
\code{sps_cv} returns two objects.
 \itemize{
   \item \code{p_value}: P-value for the null hypothesis that the the average-site ATE among non-selected sites is equal to the weighted average estimator of the average-site ATE based on site-specific ATE estimates in selected sites.
   \item \code{internal}: Objects useful for internal use of the function.
 }
}
\description{
Site-level Cross-Validation for Synthetic Purposive Sampling
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
