% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps_estimator.R
\name{sps_estimator}
\alias{sps_estimator}
\title{Synthetic Purposive Sampling Estimator for the Average-Site ATE}
\usage{
sps_estimator(
  out = NULL,
  estimates_selected = NULL,
  subgroup = NULL,
  X = NULL,
  selected_sites = NULL
)
}
\arguments{
\item{out}{Output from function \code{sps()}}

\item{estimates_selected}{data.frame with two columns: the first column represents estimates of the site-specific ATEs for the selected sites and the second column represents its corresponding standard error. The number of rows is equal to the number of the selected sites and \code{rownames(estimates_selected)} should be names of the selected sites.}

\item{subgroup}{(Optional. Default = \code{NULL}). A vector that defines subgroups to estimate the subgroup average-site ATE. The length and order of \code{subgroup} should be equal to the rows in \code{X} used to get output \code{out} with function \code{sps()}.}

\item{X}{(Optional. Use this only when sites are not selected based on \code{sps()}. Default = \code{NULL}) Site-level variables for the target population of sites. Row names should be names of sites.}

\item{selected_sites}{(Optional. Use this only when sites are not selected based on \code{sps()}. Default = \code{NULL}) Names of sites users selected. This should be a subset of rownames(X).}
}
\value{
\code{sps_estimator} returns an object of \code{sps_estimator} class.
 \itemize{
   \item \code{average_site_ATE}: An estimate of the average-site ATE and its corresponding standard error.
   \item \code{site_specific_ATE}: Estimates and standard errors for the site-specific ATEs in non-selected sites.
   \item \code{bet_se}: Estimated between-site standard errors.
 }
}
\description{
Synthetic Purposive Sampling Estimator for the Average-Site ATE
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
