% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps_map_country.R
\name{sps_map_country}
\alias{sps_map_country}
\title{Creates a world map with target population and selected countries from SPS indicated.}
\usage{
sps_map_country(out, ...)
}
\arguments{
\item{out}{Output from function \code{sps()}.}

\item{...}{Arguments passed onto \code{fuzzymatch::stringdist_join()}.}
}
\value{
\code{sps_map_country} visualizes the selection of study sites on a world map.
}
\description{
Creates a world map with target population and selected countries from SPS indicated.
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
