% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps_plot.R
\name{sps_plot}
\alias{sps_plot}
\title{Plot function for \code{sps}}
\usage{
sps_plot(out, title = NULL, columns = NULL, before_selection = FALSE, X = NULL)
}
\arguments{
\item{out}{Output from function \code{sps()}. When \code{before_selection = TRUE}, set this argument to be \code{NULL}.}

\item{title}{Title of the plot}

\item{columns}{(Optional. Default = \code{NULL}) Names of columns users want to visualize. When \code{NULL}, the function plots every column.}

\item{before_selection}{Logical (\code{TRUE} or \code{FALSE}. Default = \code{FALSE}). When \code{FALSE}, the function will compare selected and non-selected sites. When \code{TRUE}, the function simply visualizes the distribution of site-level variables in the target population of sites.}

\item{X}{(Optional. Use this only when \code{before_selection = TRUE}.) Site-level variables for the target population of sites.}
}
\value{
\code{sps_plot} visualizes the distribution of site-level variables.
}
\description{
Plot function for \code{sps}
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
