% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps_plot_est.R
\name{sps_plot_est}
\alias{sps_plot_est}
\title{Plot function for \code{sps_estimator}}
\usage{
sps_plot_est(estimates_selected, sps_est)
}
\arguments{
\item{estimates_selected}{data.frame with two columns: the first column represents estimates of the site-specific ATEs for the selected sites and the second column represents its corresponding standard error. The number of rows is equal to the number of the selected sites and \code{rownames(estimates_selected)} should be names of the selected sites.}

\item{sps_est}{Output from function \code{sps_estimator()}}
}
\value{
\code{sps_plot_est} plots the estimates and 95% confidence interval bars for site-specific ATEs and an average-site ATE.
}
\description{
Plot function for \code{sps_estimator}
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
