% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_sps.R
\name{stratify_sps}
\alias{stratify_sps}
\title{Stratification for Synthetic Purposive Sampling}
\usage{
stratify_sps(X, num_site = NULL, condition = NULL)
}
\arguments{
\item{X}{Site-level variables for the target population of sites. Row names should be names of sites. X cannot contain missing data.}

\item{num_site}{A list of two elements, e.g., \code{list("at least", 1)}. This argument specifies the number of sites that should satisfy \code{condition} specified below. The first element should be either \code{at least} or \code{at most}. The second element is integer. For example, \code{list("at least", 1)} means that we stratify SPS such that we select *at least 1* site that satisfies \code{condition} (specified below).}

\item{condition}{A list of three elements, e.g., \code{list("GDP", "larger than or equal to", 1)}. This argument specifies conditions for stratification. The first element should be a name of a site-level variable. The second element should be either \code{larger than or equal to}, \code{smaller than or equal to}, or \code{between}. The third element is a vector of length 1 or 2. When the second element is \code{between}, the third element should be a vector of two values. For example, \code{list("GDP", "larger than or equal to", 1)} means that we stratify SPS such that we select \code{num_site} sites that have *GDP larger than or equal to 1*.}
}
\value{
\code{stratify_sps} returns an object of \code{stratify_sps} class, which we supply to \code{sps()}.
 \itemize{
   \item \code{C}: A matrix on the left-hand side of linear constraints. The number of columns is the number of sites in the target population (=\code{nrow(X)}) and the number of rows is the number of constraints.
   \item \code{c0}: A vector on the right-hand side of linear constraints. The length is the number of constraints.
 }
}
\description{
Stratification for Synthetic Purposive Sampling
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
