% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_sps_estimator.R
\name{summary.sps_estimator}
\alias{summary.sps_estimator}
\title{Summary function}
\usage{
\method{summary}{sps_estimator}(object, ci = 0.95, ...)
}
\arguments{
\item{object}{the output from sps_estimator()}

\item{ci}{digits used for a table of covariate profiles. Default = 2.}

\item{...}{Other arguments.}
}
\description{
Summary function
}
