% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps.R
\name{sps}
\alias{sps}
\title{Synthetic Purposive Sampling: Site Selection for External Validity}
\usage{
sps(
  X,
  N_s,
  stratify = NULL,
  lambda = c(1, 1, 0.05),
  site_selected = NULL,
  site_unavailable = NULL,
  seed = 1234
)
}
\arguments{
\item{X}{Site-level covariates for the target population of sites}

\item{N_s}{Number of study sites to be selected}

\item{stratify}{Stratify}

\item{lambda}{A}

\item{site_selected}{A}

\item{site_unavailable}{A}

\item{seed}{Numeric. `seed` used internally. Default = 1234.}
}
\value{
\code{sps} returns an object of \code{sps} class.
 \itemize{
   \item \code{ss}: Estimated external robustness.
   \item \code{W}: Estimates of the pAMCE for all factors in each bootstrap sample.
   \item \code{obj}: Estimated values of the objective function
   \item \code{X}: X
   \item \code{Q}: Q
   \item \code{N}: N
   \item \code{N_s}: N_s
   \item \code{...}: Values for internal use.
 }
}
\description{
Synthetic Purposive Sampling: Site Selection for External Validity
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
