% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps.R
\name{sps}
\alias{sps}
\title{Synthetic Purposive Sampling: Site Selection for External Validity}
\usage{
sps(
  X,
  N_s,
  stratify = NULL,
  site_include = NULL,
  site_exclude = NULL,
  lambda = c(1, 1, 0),
  seed = 1234,
  max_iter = 10
)
}
\arguments{
\item{X}{Site-level variables for the target population of sites. Row names should be names of sites. X cannot contain missing data.}

\item{N_s}{Number of study sites to be selected.}

\item{stratify}{(Optional. Default = \code{NULL}) Output from function \code{stratify_sps()}. This argument helps users incorporate practical and logistical constraints. See examples on \url{http://naokiegami.com/spsR/articles/stratify_sps.html}}

\item{site_include}{(Optional. Default = \code{NULL}) Names of sites users want to always include (or have already selected).}

\item{site_exclude}{(Optional. Default = \code{NULL}) Names of sites users want to always exclude.}

\item{lambda}{Values of the tuning parameters. If users want to change how to balance three parts of the objective function, they can change \code{lambda}. Default values are \code{c(1, 1, 0)}. Users who want to fine-tune the tuning parameters, please see the methodological details in Egami and Lee (2023+) (https://naokiegami.com/paper/sps.pdf).}

\item{seed}{Numeric. \code{seed} used internally. Default = \code{1234}.}

\item{max_iter}{Numeric. The number of iterations used in the optimization. Default = \code{10}.}
}
\value{
\code{sps} returns an object of \code{sps} class.
 \itemize{
   \item \code{selected_sites}: Names of sites the SPS algorithm selected.
   \item \code{W}: Estimated weights to approximate non-selected sites using selected sites. \code{W} will be used in the subsequent estimation of the average-site ATE.
   \item \code{obj}: Estimated values of the objective function, separately for three parts.
   \item \code{internal}: Objects useful for internal use of the function.
 }
}
\description{
Synthetic Purposive Sampling: Site Selection for External Validity
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
